bnxz

+++ src/esercizio2/Esercizio2.java
@@ -10,17 +10,17 @@ public class Esercizio2 {
 
         Scanner scanner = new Scanner(System.in);
         System.out.println("Quanti km percorsi?");
-        double kmPercorsi = scanner.nextInt();
+        double kmPercorsi = scanner.nextDouble();
 
         System.out.println("Quanti litri consumati?");
-        double litriConsumati = scanner.nextInt();
+        double litriConsumati = scanner.nextDouble();
 
         double risultato = kmPercorsi/litriConsumati;
 
         System.out.println(risultato);
         } catch (ArithmeticException e) {
             System.out.println("impossibile dividere per 0");
-        }
+        } // il catch non prende la divisione per 0 come errore perchè la divisione per 0 esiste ed è infinity
